requirements.txt
celery[redis]
flower
python-dotenv
sqlalchemy
psycopg2
Flask
requests

